Add ApplicationController#question_set.
Returns the question set which corresponds with the current partition,
raising an error if it is not present.

Aliased question_set and partition as require_question_set! and
require_partition!, and removed load_question_set.

The QuestionSet is now loaded by ApplicationController, since all
controllers used the older load_question_set filter anyway.