Reuse the scenario ID when validation fails
If the user has JavaScript disabled, we have to fall back to validating
the user details form in ActiveRecord. However, this caused a NEW
scenario to be created, absent any of the inputs from the user's
answers. Instead we now create the ApiGateway with the original scenario
ID and the inputs and mix results are retained.

Closes #219